requirements.txt
requests2==2.16.0
lxml==4.5.0
beautifulsoup4==4.8.2
selenium==3.141.0
pandas==1.0.1